yaml_to_json_converter/requirements.txt
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.0
pytest-mock==3.15.1
python-dotenv==1.2.1
PyYAML==6.0.3
